taint/test/mochitest/test_url_search_params.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Taint URL Search Params propagation</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted");
      }

      // Check untainted
      const paramsString = "q=URLUtils.searchParams&topic=api";
      const searchParams = new URLSearchParams(paramsString);

      check_untainted(searchParams.get("q"));
      check_untainted(searchParams.get("topic"));
      check_untainted(searchParams.toString());

      // Tainted basic checks
      const taintedParamsString = "tainted=" + String.tainted("hello")+ "&untainted=goodbye";
      const taintedSearchParams = new URLSearchParams(taintedParamsString);

      check_tainted(taintedSearchParams.get("tainted"));
      check_untainted(taintedSearchParams.get("untainted"));
      check_tainted(taintedSearchParams.toString());

      taintedSearchParams.delete("tainted");
      check_untainted(taintedSearchParams.toString());


      const paramsToEscape = new URLSearchParams("q=URLUtils." + String.tainted("search<b>") + "f<o>o" + String.tainted("</b> Params") + "&"+ String.tainted("topic") + "=api");
      const us = paramsToEscape.toString();
      ok(us.taint.length == 3, "length check");
      ok(us.substring(us.taint[0].begin, us.taint[0].end) == "search%3Cb%3E", "content check");
      ok(us.substring(us.taint[1].begin, us.taint[1].end) == '%3C%2Fb%3E+Params', "content check");
      ok(us.substring(us.taint[2].begin, us.taint[2].end) == "topic", "content check");
    </script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</html>
